Lot of changes, but in the end not all that much changed. Most prominent change is that TipiWindows are no longer added to the topLevel list in the context. This was the major leaking source.
I reverted the onClose behaviour to HIDE_ON_CLOSE. This means that closing the window by clicking on the right-upper cross, it will hide the window, and the framework can show the window again by setting the visibility back to 'true'
I did update the disposeComponent method, to make sure the window has been closed.
This means that memory will still be allocated for the window (and all it refers to, which can be a lot) while it is hidden. Still, this construction is more conform to the way of the tipi, and it still prevents multiple instances of semi-disposed windows, which was ONE of the plumbing problems.