Timeout update for Navajomap.waitForResult
After the timeout expires on Navajomap.waitForResult, the thread that called wait() will resume and continue as if a response was actually received. Since this is not the case, a null Navajo response will be given, yet the Navajomap still blocks until the (async) call has actually completed.
In this change the timeout is updated to 10 minutes. If the wait() is completed but no response is present, a clear message is logged.
On the long term, we should fix.